AppConfig 추가, 테스트 케이스 작성
- 보너스 볼 중복 확인 위치 변경 (WinLottoService에서로)
- 테스트 케이스 중 2등과 3등 테스트 수정 필요(에러 발생)
- AppConfig에 맞춰서 각 파일들 생성자 추가

+++ src/main/java/lotto/domain/Match.java
@@ -8,9 +8,9 @@ public enum Match {
     FIRST(6, 2000000000);
 
     private int numOfMatches;
-    private int price;
+    private double price;
 
-    Match(int numOfMatches, int price) {
+    Match(int numOfMatches, double price) {
         this.numOfMatches = numOfMatches;
         this.price = price;
     }
@@ -19,7 +19,7 @@ public enum Match {
         return numOfMatches;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 }
